// src/app/account-management/models/document.model.ts
export class Document {
    docId: Number;
    docTypeId: Number;
    docStorageId: Number;
    docFilePath: string;
    docFileType: string;
    docExternalNumber: string;
    docAlias: string;
    docDescription: string;
    docIssueDate: any;
    docExpiryDate: any;
    parentDocId: Number;
    docStatus: Number;
    docTags: string;
    activeFlag: Number;
    createdOn: any;
    updatedOn: any;
    internalCompany: string;
}